Refactored repository
